[TESTS] refactor Authorization test class
